refactor skeleton and remove legacy views
