--- fearless/Modules/Staking/StakingMain/View/ValidationBannerView.swift
@@ -1,25 +1,0 @@
-import UIKit
-
-final class ValidationBannerView: LocalizableView {
-    @IBOutlet private var titleLabel: UILabel!
-    @IBOutlet private var detailsLabel: UILabel!
-
-    var locale = Locale.current {
-        didSet {
-            applyLocalization()
-        }
-    }
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-
-        applyLocalization()
-    }
-
-    private func applyLocalization() {
-        titleLabel.text = R.string.localizable
-            .stakingValidatorSummaryTitle(preferredLanguages: locale.rLanguages)
-        detailsLabel.text = R.string.localizable
-            .stakingValidatorSummaryDescription(preferredLanguages: locale.rLanguages)
-    }
-}
