--- fearless/Modules/Staking/StakingMain/View/ValidationBannerView.xib
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="17701" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17703"/>
-        <capability name="Named colors" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <customFonts key="customFonts">
-        <array key="sora-Bold.otf">
-            <string>sora-rc004-0417-Bold</string>
-        </array>
-        <array key="sora-Regular.otf">
-            <string>sora-rc004-0417-Regular</string>
-        </array>
-    </customFonts>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="ValidationBannerView" customModule="fearless" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="414" height="150"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="VvZ-md-2Fm" customClass="TriangularedBlurView" customModule="fearless" customModuleProvider="target">
-                    <rect key="frame" x="0.0" y="0.0" width="414" height="150"/>
-                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    <userDefinedRuntimeAttributes>
-                        <userDefinedRuntimeAttribute type="color" keyPath="_overlayFillColor">
-                            <color key="value" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        </userDefinedRuntimeAttribute>
-                        <userDefinedRuntimeAttribute type="number" keyPath="_blurStyle">
-                            <integer key="value" value="2"/>
-                        </userDefinedRuntimeAttribute>
-                    </userDefinedRuntimeAttributes>
-                </view>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Staking for validator" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="t8u-FX-VzY">
-                    <rect key="frame" x="16" y="16" width="382" height="16.5"/>
-                    <fontDescription key="fontDescription" name="sora-rc004-0417-Bold" family="sora-rc004-0417" pointSize="14"/>
-                    <color key="textColor" name="colorWhite"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="jdK-ob-vjx">
-                    <rect key="frame" x="16" y="47.5" width="382" height="78.5"/>
-                    <string key="text">In future updates, we will add staking features for the validators.
-
-#StayFearless</string>
-                    <fontDescription key="fontDescription" name="sora-rc004-0417-Regular" family="sora-rc004-0417" pointSize="12"/>
-                    <color key="textColor" name="colorTransparentText"/>
-                    <nil key="highlightedColor"/>
-                </label>
-            </subviews>
-            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-            <constraints>
-                <constraint firstItem="jdK-ob-vjx" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="16" id="HH5-O5-TZ3"/>
-                <constraint firstItem="VvZ-md-2Fm" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="VnO-QM-5rf"/>
-                <constraint firstItem="t8u-FX-VzY" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="16" id="Wmc-30-qdg"/>
-                <constraint firstAttribute="bottom" secondItem="jdK-ob-vjx" secondAttribute="bottom" constant="24" id="ZaD-va-O6U"/>
-                <constraint firstAttribute="trailing" secondItem="jdK-ob-vjx" secondAttribute="trailing" constant="16" id="bxl-h3-x50"/>
-                <constraint firstAttribute="trailing" secondItem="VvZ-md-2Fm" secondAttribute="trailing" id="cmF-Fp-gUG"/>
-                <constraint firstAttribute="trailing" secondItem="t8u-FX-VzY" secondAttribute="trailing" constant="16" id="mLa-xH-R5C"/>
-                <constraint firstAttribute="bottom" secondItem="VvZ-md-2Fm" secondAttribute="bottom" id="sCs-vR-bkP"/>
-                <constraint firstItem="t8u-FX-VzY" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="16" id="t1A-kC-LIi"/>
-                <constraint firstItem="VvZ-md-2Fm" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="wwl-oI-9QK"/>
-                <constraint firstItem="jdK-ob-vjx" firstAttribute="top" secondItem="t8u-FX-VzY" secondAttribute="bottom" constant="15" id="xHl-dL-vSZ"/>
-            </constraints>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <connections>
-                <outlet property="detailsLabel" destination="jdK-ob-vjx" id="cP5-WO-H1h"/>
-                <outlet property="titleLabel" destination="t8u-FX-VzY" id="vKm-rT-Vvz"/>
-            </connections>
-            <point key="canvasLocation" x="65" y="112"/>
-        </view>
-    </objects>
-    <resources>
-        <namedColor name="colorTransparentText">
-            <color red="1" green="1" blue="1" alpha="0.63999998569488525" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-        <namedColor name="colorWhite">
-            <color red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-    </resources>
-</document>
